Spring2024/demo/documentai_batch_processing_app/requirements.txt
google-cloud-documentai==2.26.0
google-cloud-storage==2.8.0
Pillow==9.4.0
google-cloud-translate==3.11.3
google-generativeai==0.5.2
ipython==7.34.0
google-generativeai==0.5.2